requirements.txt
termcolor
selenium
whois
pyfiglet
python-dotenv